noise field support for array return, gradient shade
