+++ shades/__init__.py
@@ -10,4 +10,4 @@ from .noise_fields import *  # noqa
 from .shades import *  # noqa
 from .utils import *  # noqa
 
-__version__ = "3.0.0"
+__version__ = "2.0.0"
